Extract heading components into directories and update file paths
