+++ client/src/components/day/heading/Day.js
@@ -1,6 +1,6 @@
 import React, {useContext, useEffect} from 'react'
-import AppContext from "../context/application/AppContext"
-import pageFlip from "../sounds/page-flip.wav"
+import AppContext from "../../../context/application/AppContext"
+import pageFlip from "../../../sounds/page-flip.wav"
 
 const Day = () => {
